podcast-summarizer/requirements.txt
fastapi
uvicorn
httpx
pydantic
sqlalchemy
pymysql
python-dotenv
boto3
faster-whisper
gtts
google-generativeai
google-genai